// Project_Apex/frontend/src/api/models/CryptoDepositAddressPublic.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CryptoDepositAddressPublic = {
    coin: string;
    network: string;
    address: string;
    memo?: string | null;
    is_active?: boolean;
    notes?: string | null;
    id: string;
    created_at: string;
    updated_at: string;
    updated_by_id?: string | null;
    updated_by_email?: string | null;
};